+++ smartcon_app/lib/controller/auth.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:smartcon_app/models/user.dart';
+import 'package:smartcon_app/model/user.dart';
 
 import 'database.dart';
 
@@ -9,7 +9,7 @@ class AuthService {
   final FirebaseAuth _auth = FirebaseAuth.instance;
 
   // create user obj based on firebase user
-  SmartconUser _userFromFirebaseUser(User user){
+  SmartconUser _userFromFirebaseUser(User user) {
     return user != null ? SmartconUser(uid: user.uid) : null;
   }
 
@@ -38,11 +38,10 @@ class AuthService {
       assert(user.uid == currentUser.uid);
 
       if (result.additionalUserInfo.isNewUser) {
-        await DatabaseService(uid: user.uid).updateProfile('',[]);
+        await DatabaseService(uid: user.uid).updateProfile('', []);
       }
       return _userFromFirebaseUser(user);
-    }
-    catch (error) {
+    } catch (error) {
       print(error.toString());
       return null;
     }
