+++ smartcon_app/lib/controller/database.dart
@@ -1,23 +1,25 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:smartcon_app/models/conference.dart';
-import 'package:smartcon_app/models/session.dart';
-import 'package:smartcon_app/models/user.dart';
+import 'package:smartcon_app/model/conference.dart';
+import 'package:smartcon_app/model/session.dart';
+import 'package:smartcon_app/model/user.dart';
 
 class DatabaseService {
-
   final String uid; // User Id
   DatabaseService({this.uid});
 
   // COLLECTIONS
-  final CollectionReference profiles = FirebaseFirestore.instance.collection("profiles");
-  final CollectionReference conferencesCollection = FirebaseFirestore.instance.collection("conferences");
-  final CollectionReference sessionSuggestions = FirebaseFirestore.instance.collection("sessionSuggestions");
+  final CollectionReference profiles =
+      FirebaseFirestore.instance.collection("profiles");
+  final CollectionReference conferencesCollection =
+      FirebaseFirestore.instance.collection("conferences");
+  final CollectionReference sessionSuggestions =
+      FirebaseFirestore.instance.collection("sessionSuggestions");
 
   // SESSION SUGGESTIONS DATA
 
   List<SessionQuestion> _conferenceQuizFromSnapshot(QuerySnapshot snapshot) {
     return snapshot.docs.map((doc) {
-      return  SessionQuestion(
+      return SessionQuestion(
         sessionId: doc.id,
         question: doc.data()['question'],
         options: List.from(doc.data()['options']),
@@ -28,27 +30,34 @@ class DatabaseService {
     }).toList();
   }
 
-  Stream<List<SessionQuestion>> getQuizQuestions(String conferenceId){
-    return conferencesCollection.doc(conferenceId).collection('sessions').snapshots().map(_conferenceQuizFromSnapshot);
+  Stream<List<SessionQuestion>> getQuizQuestions(String conferenceId) {
+    return conferencesCollection
+        .doc(conferenceId)
+        .collection('sessions')
+        .snapshots()
+        .map(_conferenceQuizFromSnapshot);
   }
 
   List<Session> _conferenceSessionsFromSnapshot(QuerySnapshot snapshot) {
     return snapshot.docs.map((doc) {
-      return  Session(
-        sessionId: doc.id,
-        name: doc.data()['name'],
-        topics: List.from(doc.data()['topics']),
-        speakers: List.from(doc.data()['speakers']),
-        begin: doc.data()['begin'].toDate().toUtc().add(Duration(hours: 1)),
-        end: doc.data()['end'].toDate().toUtc().add(Duration(hours: 1)),
-        website: doc.data()['website'],
-        description: doc.data()['description']
-      );
+      return Session(
+          sessionId: doc.id,
+          name: doc.data()['name'],
+          topics: List.from(doc.data()['topics']),
+          speakers: List.from(doc.data()['speakers']),
+          begin: doc.data()['begin'].toDate().toUtc().add(Duration(hours: 1)),
+          end: doc.data()['end'].toDate().toUtc().add(Duration(hours: 1)),
+          website: doc.data()['website'],
+          description: doc.data()['description']);
     }).toList();
   }
 
-  Stream<List<Session>> getConferenceSessions(String conferenceId){
-    return conferencesCollection.doc(conferenceId).collection('sessions').snapshots().map(_conferenceSessionsFromSnapshot);
+  Stream<List<Session>> getConferenceSessions(String conferenceId) {
+    return conferencesCollection
+        .doc(conferenceId)
+        .collection('sessions')
+        .snapshots()
+        .map(_conferenceSessionsFromSnapshot);
   }
 
   // Add User Document to Conference Suggestions Collection
@@ -57,26 +66,33 @@ class DatabaseService {
   }
 
   // Save the Session Suggestion te User got for a certain Conference
-  Future<void> addSessionSuggestion(String conferenceId, String sessionId) async{
-    return await sessionSuggestions.doc(uid).collection(conferenceId).doc(sessionId).set({});
+  Future<void> addSessionSuggestion(
+      String conferenceId, String sessionId) async {
+    return await sessionSuggestions
+        .doc(uid)
+        .collection(conferenceId)
+        .doc(sessionId)
+        .set({});
   }
 
   // Get Sessions that were suggested to the user for a certain Conference
   Future<List<String>> getSuggestedSessions(String conferenceId) async {
-    return await sessionSuggestions.doc(uid).collection(conferenceId).get().then
-      ((value) => value.docs.map((doc) {
-      return doc.id.toString();}).toList());
+    return await sessionSuggestions
+        .doc(uid)
+        .collection(conferenceId)
+        .get()
+        .then((value) => value.docs.map((doc) {
+              return doc.id.toString();
+            }).toList());
   }
-
 
   // PROFILE DATA
 
   // Update User Profile - Interests and District
   Future<void> updateProfile(String district, List<String> interests) async {
-    return await profiles.doc(uid).set({
-      "district": district,
-      "interests": interests
-    });
+    return await profiles
+        .doc(uid)
+        .set({"district": district, "interests": interests});
   }
 
   // Transform database user profile info into a instance of UserData
@@ -84,8 +100,7 @@ class DatabaseService {
     return UserData(
         uid: uid,
         district: snapshot.data()['district'],
-        interests: List.from(snapshot.data()['interests'])
-    );
+        interests: List.from(snapshot.data()['interests']));
   }
 
   // Get UserData stream
@@ -100,9 +115,12 @@ class DatabaseService {
 
   // Transform database Conference Categories into a List of category names
   List<String> _categoryListFromSnapshot(QuerySnapshot snapshot) {
-    return snapshot .docs.map((doc) {
-      return doc.data()['category'].toString();
-    }).toSet().toList();
+    return snapshot.docs
+        .map((doc) {
+          return doc.data()['category'].toString();
+        })
+        .toSet()
+        .toList();
   }
 
   // CONFERENCES DATA
@@ -119,8 +137,7 @@ class DatabaseService {
           beginDate: doc.data()['beginDate'].toDate(),
           endDate: doc.data()['endDate'].toDate(),
           website: doc.data()['website'],
-          rating: doc.data()['rating']
-      );
+          rating: doc.data()['rating']);
     }).toList();
   }
 
@@ -130,19 +147,19 @@ class DatabaseService {
   }
 
   // Insert Conference into Database
-    addConference(Conference conference) async {
+  addConference(Conference conference) async {
     var newConference = await conferencesCollection.doc();
 
-     newConference.set({
-       'name': conference.name,
-       'category': conference.category,
-       'district': conference.district,
-       'website': conference.website,
-       'description': conference.description,
-       'beginDate': conference.beginDate,
-       'endDate': conference.endDate,
-       'rating': conference.rating,
-     });
+    newConference.set({
+      'name': conference.name,
+      'category': conference.category,
+      'district': conference.district,
+      'website': conference.website,
+      'description': conference.description,
+      'beginDate': conference.beginDate,
+      'endDate': conference.endDate,
+      'rating': conference.rating,
+    });
 
     return newConference.id;
   }
@@ -151,10 +168,14 @@ class DatabaseService {
 
   // Insert Session into Database
   Future<void> addSession(String conferenceId, Session session) async {
-    return await conferencesCollection.doc(conferenceId).collection('sessions').doc().set({
+    return await conferencesCollection
+        .doc(conferenceId)
+        .collection('sessions')
+        .doc()
+        .set({
       'name': session.name,
       'speakers': session.speakers,
-      'topics' : session.topics,
+      'topics': session.topics,
       'website': session.website,
       'description': session.description,
       'begin': session.begin,
