+++ smartcon_app/lib/view/insertConference/insertConference.dart
@@ -1,7 +1,7 @@
 import 'package:date_range_picker/date_range_picker.dart' as DateRagePicker;
 import 'package:dropdown_formfield/dropdown_formfield.dart';
 import 'package:flutter/material.dart';
-import 'package:smartcon_app/models/conference.dart';
+import 'package:smartcon_app/model/conference.dart';
 import 'conferenceSessions.dart';
 
 class InsertConference extends StatefulWidget {
@@ -33,7 +33,7 @@ class InsertConferenceState extends State<InsertConference> {
     });
   }
 
-  _buildConference(){
+  _buildConference() {
     _conference = new Conference(
         name: _name,
         category: _category,
@@ -304,7 +304,7 @@ class InsertConferenceState extends State<InsertConference> {
                   Container(
                     child: Text(
                       "Insert Conference",
-                      style:TextStyle(
+                      style: TextStyle(
                         fontFamily: 'Rubik',
                         color: Color(0xFF5BBDB8),
                         fontSize: 32.0,
@@ -336,19 +336,18 @@ class InsertConferenceState extends State<InsertConference> {
                 Container(
                   padding: EdgeInsets.all(8),
                   child: RaisedButton(
-                    color: Colors.white,
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(10.0),
-                        side: BorderSide(color: Colors.black26, width: 2)),
-                    child: Text('NEXT',
-                        style: TextStyle(
-                          color: Colors.black38,
-                          fontSize: 14.0,
-                          fontWeight: FontWeight.w700,
-                          fontFamily: 'Rubik',
-                        )),
-                    onPressed: onNext
-                  ),
+                      color: Colors.white,
+                      shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(10.0),
+                          side: BorderSide(color: Colors.black26, width: 2)),
+                      child: Text('NEXT',
+                          style: TextStyle(
+                            color: Colors.black38,
+                            fontSize: 14.0,
+                            fontWeight: FontWeight.w700,
+                            fontFamily: 'Rubik',
+                          )),
+                      onPressed: onNext),
                 ),
               ],
             ),
@@ -358,15 +357,17 @@ class InsertConferenceState extends State<InsertConference> {
     );
   }
 
-  onNext(){
-      if (_formKey.currentState.validate()) {
-        _formKey.currentState.save();
-        _buildConference();
-        Navigator.push(
-          context,
-          MaterialPageRoute(
-              builder: (context) => conferenceSessions(conference: _conference,)),
-        );
-      }
+  onNext() {
+    if (_formKey.currentState.validate()) {
+      _formKey.currentState.save();
+      _buildConference();
+      Navigator.push(
+        context,
+        MaterialPageRoute(
+            builder: (context) => conferenceSessions(
+                  conference: _conference,
+                )),
+      );
+    }
   }
 }
