+++ smartcon_app/lib/view/sessionSuggestions/sessionSuggestionsCode.dart
@@ -2,10 +2,10 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:smartcon_app/models/conference.dart';
-import 'package:smartcon_app/models/user.dart';
-import 'package:smartcon_app/screens/sessionSuggestions/sessionSuggestions.dart';
-import 'package:smartcon_app/services/database.dart';
+import 'package:smartcon_app/model/conference.dart';
+import 'package:smartcon_app/model/user.dart';
+import 'package:smartcon_app/view/sessionSuggestions/sessionSuggestions.dart';
+import 'package:smartcon_app/controller/database.dart';
 
 import 'answerQuiz.dart';
 
@@ -17,7 +17,6 @@ class SessionSuggestionsCode extends StatefulWidget {
 }
 
 class SessionSuggestionsCodeState extends State<SessionSuggestionsCode> {
-
   final GlobalKey<FormState> _confCodeKey = GlobalKey<FormState>();
 
   String _code;
@@ -25,25 +24,36 @@ class SessionSuggestionsCodeState extends State<SessionSuggestionsCode> {
   Widget _conferenceCode(List<Conference> conferences) {
     final user = Provider.of<SmartconUser>(context);
 
-    String _getConferenceName(){
-      for(int i = 0; i< conferences.length; i++){
-        if(conferences[i].confId == _code)
-          return conferences[i].name;
+    String _getConferenceName() {
+      for (int i = 0; i < conferences.length; i++) {
+        if (conferences[i].confId == _code) return conferences[i].name;
       }
     }
 
-    _handleConferenceCode() async{
+    _handleConferenceCode() async {
       await DatabaseService(uid: user.uid).addUserToSessionSuggestions();
-      List<String> sessions = await DatabaseService(uid: user.uid).getSuggestedSessions(_code);
+      List<String> sessions =
+          await DatabaseService(uid: user.uid).getSuggestedSessions(_code);
       String conferenceName = _getConferenceName();
 
-      if(sessions.isEmpty){
-        Navigator.push( context, MaterialPageRoute(builder: (context) =>
-            AnswerQuiz(conferenceId: _code, conferenceName: conferenceName,)), );
-      }
-      else{
-        Navigator.push( context, MaterialPageRoute(builder: (context) =>
-            SessionSuggestions(conferenceId: _code, conferenceName: conferenceName, suggestedSessionIds: sessions)), );
+      if (sessions.isEmpty) {
+        Navigator.push(
+          context,
+          MaterialPageRoute(
+              builder: (context) => AnswerQuiz(
+                    conferenceId: _code,
+                    conferenceName: conferenceName,
+                  )),
+        );
+      } else {
+        Navigator.push(
+          context,
+          MaterialPageRoute(
+              builder: (context) => SessionSuggestions(
+                  conferenceId: _code,
+                  conferenceName: conferenceName,
+                  suggestedSessionIds: sessions)),
+        );
       }
     }
 
@@ -52,14 +62,18 @@ class SessionSuggestionsCodeState extends State<SessionSuggestionsCode> {
         child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-              Text("Please insert your ticket code to get session suggestions!\n"
+              Text(
+                  "Please insert your ticket code to get session suggestions!\n"
                   "If you haven't bought a ticket to any conference check Conference Suggestions.",
                   style: TextStyle(
-                    color:Colors.black38,
+                    color: Colors.black38,
                     fontSize: 16.0,
                     fontWeight: FontWeight.w400,
-                    fontFamily: 'Rubik',)),
-              SizedBox(height: 20,),
+                    fontFamily: 'Rubik',
+                  )),
+              SizedBox(
+                height: 20,
+              ),
               Container(
                 alignment: Alignment.topLeft,
                 child: Text("DEMO CODE: conf2",
@@ -67,9 +81,12 @@ class SessionSuggestionsCodeState extends State<SessionSuggestionsCode> {
                       color: Color(0xFF637DEB),
                       fontSize: 16.0,
                       fontWeight: FontWeight.w400,
-                      fontFamily: 'Rubik',)),
+                      fontFamily: 'Rubik',
+                    )),
               ),
-              SizedBox(height: 20,),
+              SizedBox(
+                height: 20,
+              ),
               TextFormField(
                 decoration: new InputDecoration(
                   labelText: "Code",
@@ -83,7 +100,8 @@ class SessionSuggestionsCodeState extends State<SessionSuggestionsCode> {
                 validator: (String value) {
                   if (value.isEmpty)
                     return 'Code is Required!';
-                  else if(!conferences.contains(new Conference.onlyId(confId: _code))){
+                  else if (!conferences
+                      .contains(new Conference.onlyId(confId: _code))) {
                     return 'Invalid Ticket!';
                   }
                   return null;
@@ -101,33 +119,31 @@ class SessionSuggestionsCodeState extends State<SessionSuggestionsCode> {
                       side: BorderSide(color: Colors.black26, width: 2)),
                   child: Text('GET SUGGESTIONS',
                       style: TextStyle(
-                        color:Colors.black38,
+                        color: Colors.black38,
                         fontSize: 14.0,
                         fontWeight: FontWeight.w700,
-                        fontFamily: 'Rubik',)),
+                        fontFamily: 'Rubik',
+                      )),
                   onPressed: () async {
                     _confCodeKey.currentState.save();
-                    if (!_confCodeKey.currentState.validate())
-                      return;
+                    if (!_confCodeKey.currentState.validate()) return;
 
                     _handleConferenceCode();
                     //Send to API
                   },
                 ),
               ),
-            ])
-    );
+            ]));
   }
 
   @override
   Widget build(BuildContext context) {
-
     final user = Provider.of<SmartconUser>(context);
 
     return StreamBuilder<List<Conference>>(
         stream: DatabaseService(uid: user.uid).conferences,
         builder: (context, snapshot) {
-          if(snapshot.hasData){
+          if (snapshot.hasData) {
             List<Conference> conferences = snapshot.data;
 
             return Scaffold(
@@ -136,7 +152,8 @@ class SessionSuggestionsCodeState extends State<SessionSuggestionsCode> {
                   Row(children: <Widget>[
                     Container(
                       width: MediaQuery.of(context).size.width,
-                      child: Image.asset('images/pageHeader.png', fit: BoxFit.fill),
+                      child: Image.asset('images/pageHeader.png',
+                          fit: BoxFit.fill),
                     ),
                   ]),
                   Padding(
@@ -163,9 +180,7 @@ class SessionSuggestionsCodeState extends State<SessionSuggestionsCode> {
                 ]),
               ),
             );
-          }
-
-          else{
+          } else {
             return Container();
           }
         });
